Lector_Python/requirements.txt
pyserial
pymysql
paramiko
python-dotenv
sshtunnel
pytz
